Move Pie Chart to new file and start on language list
